requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
requests==2.31.0
gTTS==2.5.1
python-telegram-bot==13.15
Werkzeug==2.3.7
email-validator==2.1.0.post1
